#59 Get rid of *Markup classes for lexemes and word forms
Moved the responsibility to access ANTLR data to problem domain classes.
