+++ src/main/java/org/tendiwa/lexeme/ParsedWordForm.java
@@ -3,6 +3,8 @@ package org.tendiwa.lexeme;
 
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
+import org.antlr.v4.runtime.tree.TerminalNode;
+import org.tendiwa.lexeme.antlr.WordBundleParser;
 import org.tenidwa.collections.utils.Collectors;
 
 /**
@@ -10,31 +12,32 @@ import org.tenidwa.collections.utils.Collectors;
  * @version $Id$
  * @since 0.1
  */
-public final class WordFormFromMarkup implements WordForm {
-    private final WordFormMarkup markup;
+public final class ParsedWordForm implements WordForm {
     private final Grammar grammar;
+    private final WordBundleParser.EntryContext ctx;
 
-    WordFormFromMarkup(WordFormMarkup markup, Grammar grammar) {
-        this.markup = markup;
+    ParsedWordForm(Grammar grammar, WordBundleParser.EntryContext ctx) {
         this.grammar = grammar;
+        this.ctx = ctx;
     }
 
     @Override
     public String spelling() {
-        return this.markup.spelling();
+        return this.ctx.WORD_FORM().getText();
     }
 
     @Override
     public int similarity(ImmutableSet<Grammeme> grammemes) {
         return Sets.intersection(
-            this.grammaticalMeaning(),
-            grammemes
+            grammemes,
+            this.grammemes()
         ).size();
     }
 
-    private ImmutableSet<Grammeme> grammaticalMeaning() {
-        return this.markup.grammemes()
+    private ImmutableSet<Grammeme> grammemes() {
+        return this.ctx.grammemes().GRAMMEME()
             .stream()
+            .map(TerminalNode::getText)
             .map(this.grammar::grammemeByName)
             .collect(Collectors.toImmutableSet());
     }
