--- src/test/java/org/tendiwa/lexeme/ParsedWordFormMarkupTest.java
@@ -1,53 +1,0 @@
-package org.tendiwa.lexeme;
-
-import java.io.IOException;
-import org.antlr.v4.runtime.ANTLRInputStream;
-import org.antlr.v4.runtime.CommonTokenStream;
-import org.apache.commons.io.IOUtils;
-import org.hamcrest.CoreMatchers;
-import org.hamcrest.MatcherAssert;
-import org.junit.Test;
-import org.tendiwa.lexeme.antlr.WordBundleLexer;
-import org.tendiwa.lexeme.antlr.WordBundleParser;
-
-/**
- * @since 0.1
- */
-public final class ParsedWordFormMarkupTest {
-    @Test
-    public void seesAllComponents() throws Exception {
-        final WordFormMarkup markup =
-            new ParsedWordFormMarkup(
-                new WordBundleParser(
-                    new CommonTokenStream(
-                        this.lexerInLexemesMode("dudes [Plur]")
-                    )
-                ).entry()
-            );
-        MatcherAssert.assertThat(
-            markup.spelling(),
-            CoreMatchers.is("dudes")
-        );
-        MatcherAssert.assertThat(
-            markup.grammemes().contains("Plur"),
-            CoreMatchers.is(true)
-        );
-    }
-
-    /**
-     * Constructs a lexer that starts in {@link WordBundleLexer#LEXEMES} mode.
-     * @param input Text to parse.
-     * @return Lexer in {@link WordBundleLexer#LEXEMES} mode.
-     */
-    private WordBundleLexer lexerInLexemesMode(String input)
-        throws IOException {
-        final WordBundleLexer lexer = new WordBundleLexer(
-            new ANTLRInputStream(
-                IOUtils.toInputStream(input)
-            )
-        );
-        lexer.mode(WordBundleLexer.LEXEMES);
-        return lexer;
-    }
-
-}
